Débug de création de la console de jeux : méthode configure() dans GestionnaireDeJeu

+++ Bateau.cpp
@@ -50,7 +50,7 @@ bool Bateau::getBateauState()
 {
     return hitPoints;
 }
-void Bateau::setBateauSize(unsigned int size)
+void Bateau::setBateauSize(uint8_t size)
 {
     bateauSize = size;
 }
